requirements.txt
fastapi
uvicorn
requests
unstructured
sentence-transformers
faiss-cpu
openai
python-multipart
pydantic
transformers